Edited TrackingClass skeleton
This commit adds a part of relation with the DetectionClass and adds a
varaible to hold the vector of detected obstacles.
Added parameters to the constructor to initialise the detection model in
the DetectionClass.
Merged the track() function with the assignID() function, changed the
return type and added a function parameter. The function will now assign
ID to each bounding box and compare the new detections for tracking.
Changed distFromCamera() and distFromCar() return type to return (x,y,z)
distance of the object.
Added extra Doxygen documentation.